refactor: move types to seprate file
