+++ src/context.ts
@@ -1,60 +1,5 @@
-import type { RequestListener } from 'http'
 import defu from 'defu'
-import { NuxtConfig, NuxtOptions } from '@nuxt/types'
-import type { Listener } from 'listhen'
-import type { Browser, LaunchOptions } from 'playwright'
-
-export interface NuxtTestOptions {
-  rootDir: string
-  configFile: string
-  config: NuxtConfig
-
-  randomBuildDir: boolean
-  randomPort: boolean
-
-  build: boolean
-  generate: boolean
-
-  setupTimeout: number
-  waitFor: number
-
-  browser: boolean
-  browserOptions: {
-    type: 'chromium' | 'firefox' | 'webkit'
-    launch?: LaunchOptions
-  }
-
-  server: boolean
-}
-
-export interface Nuxt {
-  server: {
-    app: RequestListener,
-  },
-  options: NuxtOptions
-  ready: () => any
-  close: (callback?: Function) => any
-  resolver: any
-  moduleContainer: any
-  resolveAlias(path: string): string
-  resolvePath(path: string, opts?: any): string
-  renderRoute(...args: any[]): any
-  renderAndGetWindow(url: string, opts?: any, config?: any): any
-}
-
-export interface NuxtTestContext {
-  options: Partial<NuxtTestOptions>
-
-  nuxt?: Nuxt
-
-  builder?: {
-    build: () => any
-  }
-
-  listener?: Listener
-
-  browser?: Browser
-}
+import { NuxtTestContext, NuxtTestOptions } from './types'
 
 let currentContext: NuxtTestContext
