+++ src/index.ts
@@ -7,3 +7,4 @@ export * from './mocks'
 export * from './nuxt'
 export * from './server'
 export * from './setup'
+export * from './types'
